Add Number, Price, Total

+++ src/main/java/com/PWS/Friday/ProductServiceController.java
@@ -23,41 +23,58 @@ import org.springframework.web.bind.annotation.RestController;
  */
 @RestController
 public class ProductServiceController {
-   private static Map<String, Product> productRepo = new HashMap<>();
-   static {
-      Product honey = new Product();
-      honey.setId("1");
-      honey.setName("Honey");
-      productRepo.put(honey.getId(), honey);
-      
-      Product almond = new Product();
-      almond.setId("2");
-      almond.setName("Almond");
-      productRepo.put(almond.getId(), almond);
-   }
+    private static Map<String, Product> productRepo = new HashMap<>();
+    static {
+        Product honey = new Product();
+        honey.setId("1");
+        honey.setName("Honey");
+        honey.setNumber(2);
+        honey.setPrice(10000);
+        honey.setTotal();
+        productRepo.put(honey.getId(), honey);
+        
+        Product almond = new Product();
+        almond.setId("3");
+        almond.setName("Almond");
+        almond.setNumber(3);
+        almond.setPrice(20000);
+        almond.setTotal();
+        productRepo.put(almond.getId(), almond);
+    }
+    
+    @RequestMapping(value = "/products/{id}", method = RequestMethod.DELETE)
+    public ResponseEntity<Object> delete(@PathVariable("id") String id) {
+        productRepo.remove(id);
+        return new ResponseEntity<>("Product is deleted successsfully", HttpStatus.OK);
+    }
    
-   @RequestMapping(value = "/products/{id}", method = RequestMethod.DELETE)
-   public ResponseEntity<Object> delete(@PathVariable("id") String id) { 
-      productRepo.remove(id);
-      return new ResponseEntity<>("Product is deleted successsfully", HttpStatus.OK);
-   }
-   
-   @RequestMapping(value = "/products/{id}", method = RequestMethod.PUT)
-   public ResponseEntity<Object> updateProduct(@PathVariable("id") String id, @RequestBody Product product) { 
-      productRepo.remove(id);
-      product.setId(id);
-      productRepo.put(id, product);
-      return new ResponseEntity<>("Product is updated successsfully", HttpStatus.OK);
-   }
-   
-   @RequestMapping(value = "/products", method = RequestMethod.POST)
-   public ResponseEntity<Object> createProduct(@RequestBody Product product) {
-      productRepo.put(product.getId(), product);
-      return new ResponseEntity<>("Product is created successfully", HttpStatus.CREATED);
-   }
-   
-   @RequestMapping(value = "/products")
-   public ResponseEntity<Object> getProduct() {
-      return new ResponseEntity<>(productRepo.values(), HttpStatus.OK);
-   }
+    @RequestMapping(value = "/products/{id}", method = RequestMethod.PUT)
+    public ResponseEntity<Object> updateProduct(@PathVariable("id") String id, @RequestBody Product product){
+
+        if(!productRepo.containsKey(id)){
+            return new ResponseEntity<>("Product Not Found, Please check again", HttpStatus.NOT_FOUND);
+        }
+        else{
+            productRepo.remove(id);
+            product.setId(id);
+            productRepo.put(id, product);
+            return new  ResponseEntity<>("Product is updated Successfully",HttpStatus.OK);
+        }
+    }
+  
+    @RequestMapping(value = "/products", method = RequestMethod.POST)
+    public ResponseEntity<Object> createProduct(@RequestBody Product product){
+
+        if(productRepo.containsKey(product.getId())){ 
+            return new ResponseEntity<>("ID Product Cannot be the Same, please check again", HttpStatus.OK);
+        }
+        else{
+            productRepo.put(product.getId(), product);
+            return new ResponseEntity<>("Product is created Successfully", HttpStatus.CREATED);
+        }
+    }
+    @RequestMapping(value = "/products")
+    public ResponseEntity<Object> getproduct() {
+        return new ResponseEntity<>(productRepo.values(), HttpStatus. OK);
+    }
 }
